wip restructure pools module
