+++ src/pages/pools/pod-staking.jsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 
-import { PoolsTabs } from "@/components/pages/pools/PoolsTabs";
+import { PoolsTabs } from "@/src/modules/pools/bond/PoolsTabs";
 import { PodStakingPage } from "@/components/pages/pools/pod-staking";
 import { ComingSoon } from "@/components/pages/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
