+++ src/pages/pools/staking.jsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 
-import { PoolsTabs } from "@/components/pages/pools/PoolsTabs";
+import { PoolsTabs } from "@/src/modules/pools/bond/PoolsTabs";
 import { StakingPage } from "@/components/pages/pools/staking";
 import { isFeatureEnabled } from "@/src/config/environment";
 import { ComingSoon } from "@/components/pages/ComingSoon";
